e-canteen-backend/pyproject.toml
[tool.poetry]
name = "e-canteen-backend"
version = "0.1.0"
description = ""
authors = ["KarthikNair45 <karthiknair4569@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
marshmallow = "^3.23.1"
flask-marshmallow = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
